simplify the JS api
